More documentation, remove commented code, some requested changes
